requirements.txt
aniso8601==8.0.0
attrs==20.2.0
click==7.1.2
Flask==1.1.1
flask-restx==0.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
pyrsistent==0.17.3
pytz==2020.1
six==1.15.0
Werkzeug==0.16.1
Flask-SQLAlchemy==2.4.1
psycopg2-binary==2.8.4
pytest==5.3.5
# gunicorn==20.0.4
pytest-cov==2.8.1
flake8===3.7.9
black==19.10b0
isort==4.3.21
# createsend
requests
# grequests
